[TechDraw] Extract LineFormat
