+++ src/Mod/TechDraw/App/CenterLine.cpp
@@ -40,6 +40,7 @@
 #include "DrawViewPart.h"
 #include "Geometry.h"
 #include "GeometryObject.h"
+#include "LineFormat.h"
 #include "CenterLinePy.h"
 
 using namespace TechDraw;
