+++ src/Mod/TechDraw/App/LineFormat.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- *   Copyright (c) 2017 WandererFan <wandererfan@gmail.com>                *
+ *   Copyright (c) 2019 WandererFan <wandererfan@gmail.com>                *
  *                                                                         *
  *   This file is part of the FreeCAD CAx development system.              *
  *                                                                         *
@@ -20,55 +20,37 @@
  *                                                                         *
  ***************************************************************************/
 
-//! LineGroup - Classes related to processing LineGroup definition CSV files
+#ifndef TECHDRAW_LINEFORMAT_H
+#define TECHDRAW_LINEFORMAT_H
 
-#ifndef TechDraw_LINEGROUP_H_
-#define TechDraw_LINEGROUP_H_
+#include <App/Color.h>
+#include <Base/Persistence.h>
 
-#include <Mod/TechDraw/TechDrawGlobal.h>
+namespace TechDraw {
 
-#include <string>
-#include <vector>
-
-namespace TechDraw
-{
-
-class TechDrawExport LineGroup
+//general purpose line format specifier
+class TechDrawExport LineFormat //: Base::Persistence
 {
 public:
-    LineGroup();
-    LineGroup(std::string groupName);
-    ~LineGroup();
-    double getWeight(std::string s);
-    void setWeight(std::string s, double weight);
-//    void setWeight(const char* s, double weight);
+    LineFormat();
+    LineFormat(int style,
+               double weight,
+               App::Color color,
+               bool visible);
+
+    int m_style;
+    double m_weight;
+    App::Color m_color;
+    bool m_visible;
+
+    static double getDefEdgeWidth();
+    static App::Color getDefEdgeColor();
+    static int getDefEdgeStyle();
+
     void dump(const char* title);
-    std::string getName(void) { return m_name; }
-    void setName(std::string s) { m_name = s; }
-
-    //static support function: split comma separated string of values into vector of numbers
-    static std::vector<double> split(std::string line);
-
-    //static support function: find group defn in file
-    static std::string getRecordFromFile(std::string parmFile, int groupNumber);
-
-    //static LineGroup maker
-    static LineGroup* lineGroupFactory(int groupNumber);
-
-    static double getDefaultWidth(std::string weightName, int groupNumber = -1);
-
-    static std::string getGroupNamesFromFile(std::string FileName);
-
-
-protected:
-    void init(void);
-
-    std::string m_name;
-    double      m_thin;
-    double      m_graphic;
-    double      m_thick;
-    double      m_extra;
+    std::string toString() const;
 };
 
 }
+
 #endif
